refactor: Merge & rename javaee modules
- merge module operaton-ejb-client-jakarta to operaton-ejb-client
- rename directory /javaee -> /jakarta-ee
- change groupId org.operaton.bpm.javaee -> org.operaton.bpm.jakartaee
- cleanup Sonar findings
