+++ bom/internal-dependencies/pom.xml
@@ -501,6 +501,21 @@
         <version>4.3.0</version>
       </dependency>
 
+      <!-- jobexecutor-ra -->
+      <dependency>
+        <groupId>org.jboss.ironjacamar</groupId>
+        <artifactId>ironjacamar-spec-api</artifactId>
+        <version>1.0.10.Final</version>
+        <scope>provided</scope>
+      </dependency>
+
+      <dependency>
+        <groupId>org.codehaus.fabric3.api</groupId>
+        <artifactId>commonj</artifactId>
+        <version>1.1.0</version>
+        <scope>provided</scope>
+      </dependency>
+
     </dependencies>
   </dependencyManagement>
 </project>
