+++ bom/operaton-only-bom/pom.xml
@@ -83,13 +83,8 @@
         <version>${project.version}</version>
       </dependency>
       <dependency>
-        <groupId>org.operaton.bpm.javaee</groupId>
+        <groupId>org.operaton.bpm.jakartaee</groupId>
         <artifactId>operaton-ejb-client</artifactId>
-        <version>${project.version}</version>
-      </dependency>
-      <dependency>
-        <groupId>org.operaton.bpm.javaee</groupId>
-        <artifactId>operaton-ejb-client-jakarta</artifactId>
         <version>${project.version}</version>
       </dependency>
       <dependency>
