+++ distro/wildfly/assembly/pom.xml
@@ -33,8 +33,8 @@
     </dependency>
 
     <dependency>
-      <groupId>org.operaton.bpm.javaee</groupId>
-      <artifactId>operaton-ejb-client-jakarta</artifactId>
+      <groupId>org.operaton.bpm.jakartaee</groupId>
+      <artifactId>operaton-ejb-client</artifactId>
     </dependency>
 
     <dependency>
