+++ jakarta-ee/ejb-client/src/main/resources/META-INF/ejb-jar.xml
