+++ jakarta-ee/ejb-client/src/main/resources/WEB-INF/ejb-jar.xml
