+++ jakarta-ee/jobexecutor-ra/src/main/java/org/operaton/bpm/container/impl/threading/ra/commonj/CommonjWorkRunnableAdapter.java
@@ -1,0 +1,50 @@
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH
+ * under one or more contributor license agreements. See the NOTICE file
+ * distributed with this work for additional information regarding copyright
+ * ownership. Camunda licenses this file to you under the Apache License,
+ * Version 2.0; you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.operaton.bpm.container.impl.threading.ra.commonj;
+
+import commonj.work.Work;
+
+/**
+ * An adapter for wrapping a Runnable as a CommonJ {@link Work} instance.
+ * 
+ * @author Daniel Meyer
+ * 
+ */
+public class CommonjWorkRunnableAdapter implements Work {
+
+  private final Runnable delegate;
+  
+  public CommonjWorkRunnableAdapter(Runnable delegate) {
+    this.delegate = delegate;
+  }
+
+  @Override
+  public void run() {
+    delegate.run();    
+  }
+
+  @Override
+  public boolean isDaemon() {
+    return false;
+  }
+
+  @Override
+  public void release() {
+    // nothing to do
+  }
+
+}
