+++ jakarta-ee/jobexecutor-ra/src/main/java/org/operaton/bpm/container/impl/threading/ra/outbound/JcaExecutorServiceManagedConnection.java
