--- javaee/ejb-client/src/main/resources/META-INF/web-fragment.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<web-fragment xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-              xmlns="http://java.sun.com/xml/ns/javaee" xmlns:web="http://java.sun.com/xml/ns/javaee/web-app_2_5.xsd"
-              xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-fragment_3_0.xsd"
-              version="3.0">
-  <listener>
-    <listener-class>org.operaton.bpm.application.impl.web.ProcessArchiveServletContextListener</listener-class>
-  </listener>
-
-  <distributable/>
-</web-fragment>
