--- javaee/jobexecutor-rar/pom.xml
@@ -1,45 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <groupId>org.operaton.bpm.javaee</groupId>
-  <artifactId>operaton-jobexecutor-rar</artifactId>
-  <packaging>rar</packaging>
-  <name>Operaton - Job Executor RAR</name>
-
-  <parent>
-    <groupId>org.operaton.bpm</groupId>
-    <artifactId>operaton-parent</artifactId>
-    <relativePath>../../parent</relativePath>
-    <version>1.0.0-beta-4-SNAPSHOT</version>
-  </parent>
-
-  <dependencyManagement>
-    <dependencies>
-      <dependency>
-        <groupId>org.operaton.bpm</groupId>
-        <artifactId>operaton-core-internal-dependencies</artifactId>
-        <version>${project.version}</version>
-        <scope>import</scope>
-        <type>pom</type>
-      </dependency>
-    </dependencies>
-  </dependencyManagement>
-
-  <dependencies>
-    <dependency>
-      <groupId>org.operaton.bpm.javaee</groupId>
-      <artifactId>operaton-jobexecutor-ra</artifactId>
-      <version>${project.version}</version>
-    </dependency>
-  </dependencies>
-
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-rar-plugin</artifactId>
-      </plugin>
-    </plugins>
-  </build>
-
-  <description>${project.name}</description>
-</project>
